### 2024-05-06 #6702 @kanishkan91
@crvernon , @biplovbhandari It seems the pdf still does not seem to generate correctly. Not sure if this is a JOSS software error or an edit to the paper that caused it.  We would need to get this fixed before it goes to review.
