### 2024-05-07 #6702 @kanishkan91
@jtagusari Would you or any of your students be interested in reviewing this manuscript for JOSS? Thought you might be a good fit given your recent JOSS publication related to the QGIS plugin.
