### 2024-05-07 #6729 @editorialbot
[REVIEW]: servir-aces: A Python Package for Training Machine Learning Models for Remote Sensing Applications
**Submitting author:** @BiplovBhandari (<a href="http://orcid.org/0000-0001-6169-8236">Biplov Bhandari</a>)
**Repository:** https://github.com/SERVIR/servir-aces
**Branch with paper.md** (empty if default branch):
**Version:** v0.0.16
