### 2024-05-20 SERVIR/servir-aces#37 @nagellette
I'm having the same issue while testing for review: https://github.com/openjournals/joss-reviews/issues/6729. The reason probably the failure to while defining the default paths in sample `config.env`. There should be a definition to `BASEDIR` as demonstrated in https://github.com/SERVIR/servir-aces/blob/main/notebook/count_sample_size.ipynb

This portion of the README.md should be improved to a flowlessly working version.

**Environment Information**

- servir-cases version: 0.0.16
- Python version: 3.10.12
- Operating System: Ubuntu 22.04.4 LTS

The package was installed from GitHub, using pip.
