### 2024-05-20 #6729 @nagellette
@biplovbhandari although there are multiple applications with GEE and Tensorflow in the context of agricultural applications that are mentioned in statement of need section I couldn't find anything actually referring to similar software and comparing yours against to them. One example I can think of is [torchgeo](https://github.com/microsoft/torchgeo) which is providing access to different sources (incl. Sentinel and Landsat which I think are also available via GEE) while providing processing and training capabilities in Pytorch. Another example would be [raster-vision](https://github.com/azavea/raster-vision). I am not aware anything in Tensorflow sourcing from GEE though I suggest looking in to similar software, referring to them while emphasizing servir-aces's contribution.

- torchgeo: https://dl.acm.org/doi/10.1145/3557915.3560953
- raster-vision: https://github.com/azavea/raster-vision
