+++ paper/paper.bib
@@ -70,3 +70,30 @@
   year={2021},
   publisher={MDPI}
 }
+
+@inproceedings{Stewart_TorchGeo_Deep_Learning_2022,
+  address={Seattle, Washington},
+  author={Stewart, Adam J. and Robinson, Caleb and Corley, Isaac A. and Ortiz, Anthony and Lavista Ferres, Juan M. and Banerjee, Arindam},
+  booktitle={Proceedings of the 30th International Conference on Advances in Geographic Information Systems},
+  doi={10.1145/3557915.3560953},
+  month=nov,
+  pages={1--12},
+  publisher={Association for Computing Machinery},
+  series={SIGSPATIAL '22},
+  title={{TorchGeo}: Deep Learning With Geospatial Data},
+  url={https://dl.acm.org/doi/10.1145/3557915.3560953},
+  year={2022}
+}
+
+@article{Wu2020,
+  doi={10.21105/joss.02305},
+  url={https://doi.org/10.21105/joss.02305},
+  year={2020},
+  publisher={The Open Journal},
+  volume={5},
+  number={51},
+  pages={2305},
+  author={Qiusheng Wu},
+  title={geemap: A Python package for interactive mapping with Google Earth Engine},
+  journal={Journal of Open Source Software}
+}
