### 2024-06-10 #6729 @kanishkan91
While the author responds, @jtagusari could you provide a quick update on your review. I saw that you had checked off a few items on your checklist. Is it that you will check off the remaining once the author makes the adjustments you suggested? Do let me know if you need anything at all from me or the author. Thank you again !
