+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "servir-aces"
-version = "0.0.16"
+version = "0.0.17"
 dynamic = [
     "dependencies",
 ]
@@ -47,7 +47,7 @@ universal = true
 
 
 [tool.bumpversion]
-current_version = "0.0.16"
+current_version = "0.0.17"
 commit = true
 tag = true
 
