### 2024-06-10 SERVIR/servir-aces#37 @biplovbhandari
@jtagusari and @nagellette Sorry for coming late in this, and thank you so much for reviewing the software package. As mentioned by both of you, I have updated the `README` as well as the `Usage` docs. In addition, I have linked notebooks in both of them so user can look into more details. This should help you to run the package successfully. Let me know how it goes.

@nagellette based on your comment on emphasizing servir-aces's contribution especially in the TF and GEE community, I have updated the Statement of Need section.
