### 2024-06-11 SERVIR/servir-aces#37 @jtagusari
@biplovbhandari

Thanks for the update.
While I have not finished my test, the following command is for Colab and should be modified if it is for local environments.

```
cd DATADIR
!gsutil -m cp -r gs://dl-book/chapter-1/* DATADIR
```

I think it should be:
```
mkdir DATADIR
gsutil -m cp -r gs://dl-book/chapter-1/* DATADIR
```

In addition, is it difficult to use smaller test data to check the functionality of the package? It is more than 16 GB and requires a lot of memory and computation. It would be great if you could prepare lightweight test cases for quick checking.

Thanks
