+++ paper/paper.md
@@ -33,7 +33,7 @@ bibliography: paper.bib
 
 Despite robust platforms, specialized technical knowledge is required to setup and run various ML/DL models, leading many practitioners, scientists, and domain experts to find it difficult to implement them. The **`servir-aces`** Python package is created to fill this gap. **`servir-aces`** siginicantly lowers the barrier for users to export training data and both train and run DL models using cloud-based technology with their GEE workflows. Several examples are provided via a runnable notebook to make it easier for scientists utilize this emerging field of DL.
 
-With petabytes of data available via GEE, and integration of the TensorFlow (TF) platfrom, models trained in TF can be easily loaded into GEE. This package provides functionalities for 1) data processing; 2) data loading from GEE; 3) feature extraction, 4) model training, and 5) model inference. The combination of TF and GEE has enabled several large scale ML and DL Remote Sensing applications. Some of them include Wetland Area Mapping [@Bakkestuen2023], Crop Type Mapping [@Poortinga2021], Surface Water Mapping [@Mayer2021], and Urban Mapping [@Parekh2021]. However, these applications tend to be developed ad-hoc without using a common library and require a very specialized domain as well as technical knowledge. In addition, several unified libraries like torchgeo [@Stewart_TorchGeo_Deep_Learning_2022] and [rastervision](https://github.com/azavea/raster-vision) exists but those are mostly targeted for pytorch user community. Some efforts for GEE & TensorFlow users stemmed out of geemap [@Wu2020] but that is mostly for tree based approaches like Random Forest, while [geospatial-ml](https://github.com/opengeos/geospatial-ml) has not seen much development since its inception. Thus there is a need for a unified libraries to train deep learning models within the GEE & TensorFlow user community. The **`servir-aces`** is a first step for that. Although this was originally deeloped for agricultural related applications, the library has matured enough to work for any kind of image segmentation tasks.
+With petabytes of data available via GEE, and integration of the TensorFlow (TF) platfrom, models trained in TF can be easily loaded into GEE. This package provides functionalities for 1) data processing; 2) data loading from GEE; 3) feature extraction, 4) model training, and 5) model inference. The combination of TF and GEE has enabled several large scale ML and DL Remote Sensing applications. Some of them include Wetland Area Mapping [@Bakkestuen2023], Crop Type Mapping [@Poortinga2021], Surface Water Mapping [@Mayer2021], and Urban Mapping [@Parekh2021]. However, these applications tend to be developed ad-hoc without using a common library and require a very specialized domain as well as technical knowledge. In addition, several unified libraries like torchgeo [@Stewart_TorchGeo_Deep_Learning_2022] and [rastervision](https://github.com/azavea/raster-vision) exists but those are mostly targeted for pytorch user community. Some efforts for GEE & TensorFlow users stemmed out of geemap [@Wu2020] but that is mostly for tree based approaches like Random Forest, while [geospatial-ml](https://github.com/opengeos/geospatial-ml) has not seen much development since its inception. Thus there is a need for a unified libraries to train deep learning models within the GEE & TensorFlow user community. The **`servir-aces`** package enables the streamlined and systematic approach and innovates by 1) developing sets of commonly employed architectures (CNN, DNN, and U-NET), 2) rapidly adjust sets of parameters for these common architectures (activation functions, optimizers, loss functions, early stopping, dropout rate, batch size etc.), 3) More effectively connect the Google ecosystem between the Google Cloud, improved methods of parallelization  via Apache beam,  of Vertex AI, and Google Earth Engine, 4) This package’s utility function allows for broader development 5) Collate sets of evaluation metrics for easier model performance comparisons . Although this was originally developed for agricultural related applications, the library has matured enough to work for any kind of image segmentation tasks.
 
 # **`servir-aces`** Audience
 
