### 2024-06-12 ba7d045
improve readme and bugfix notebooks
