+++ README.md
@@ -35,21 +35,21 @@ For quickly running this, we have already prepared and exported the training dat
 *Note: If you're looking to produce your own datasets, you can follow this [notebook](https://colab.research.google.com/drive/1LCFLeSCu969wIW8TD68-j4hfIu7WiRIK?usp=sharing) which was used to produce these training, testing, and validation datasets provided here.*
 
 ```sh
-cd DATADIR
-!gsutil -m cp -r gs://dl-book/chapter-1/* DATADIR
+mkdir DATADIR
+gsutil -m cp -r gs://dl-book/chapter-1/dnn_planet_wo_indices/* DATADIR
 ```
 
-The folder has several dataset inside it. If you're running a U-Net model, you can use the sub-folder `unet_256x256_planet_wo_indices`; for DNN model, you can use the sub-folder `dnn_planet_wo_indices`.
+The parent folder (if you run `gsutil -m cp -r gs://dl-book/chapter-1/* DATADIR`) has several dataset inside it. We use `dnn_planet_wo_indices` here because it is lightweight data and would be much faster to run. If you want to test U-Net model, you can use `unet_256x256_planet_wo_indices` folder instead. Each of these have training, testing, and validation sub-folder inside them.
 
 The datadir is constructed as `DATADIR = BASEDIR + DATADIR` \
 OUTPUT_DIR (str): The base directory where the output would be saved.
 
-You can then change the `MODEL_TYPE` which you are running. The default is "unet". You can change to "dnn" if you are using that model.
+You can then change the `MODEL_TYPE` which you are running. The default is "unet", so we need to change it to "dnn" since we will be training using that model.
 
 *Note current version does not expose all the model intracacies through the environment file but future version may include those depending on the need.*
 
 ```
-MODEL_TYPE = "unet"
+MODEL_TYPE = "dnn"
 ```
 
 Next define the `FEATURES` and the `LABELS` variables. This dataset was prepared for the rice mapping application that uses before and during growing season information. So for this dataset, here are the example `FEATURES`. Note the `FEATURES` should be in the same format as shown below
@@ -65,13 +65,13 @@ blue_during
 nir_during"
 ```
 
-Similarly, since the dataset has a single label, it is going to be
+Similarly, since the dataset has a single label, it is going to be as below. If you have mutiple labels, be sure to format them as the `FEATURES` above.
 
 ```
-LABELS = ["class"]
+LABELS = "class"
 ```
 
-The `PATCH_SHAPE` for the training datasets needs to be changed as well:
+While the DNN would not need this setting, if you want to run U-Net model, the `PATCH_SHAPE` for the training datasets needs to be changed as well:
 
 ```
 PATCH_SHAPE = (256, 256)
@@ -97,16 +97,211 @@ After all the settings been correctly setup, here's an example of how to use the
 from aces.config import Config
 from aces.model_trainer import ModelTrainer
 
-if __name__ == "__main__":
-    config_file = "config.env"
-    config = Config(config_file)
-    trainer = ModelTrainer(config)
-    trainer.train_model()
+config_file = "config.env"
+config = Config(config_file)
+trainer = ModelTrainer(config)
+trainer.train_model()
 ```
 
 Once the model is finished running, it saves the trained model, evaluate the results on the testing dataset and saves it, produces the plots and saves it, and saves any other needed item on the `MODEL_DIR`, which is constructed as `OUTPUT_DIR + MODEL_DIR_NAME`.
 
[…]
