### 2024-06-12 SERVIR/servir-aces#37 @biplovbhandari
@jtagusari Thanks for your comments. I have fixed this issue. For now, the README would walk you through running the DNN model which has very light datasets for you to test with.
For the other comments on the paper, we are still working on the statement of need and will update you once it's ready to review. Thanks!
