### 2024-06-12 39ee074
Update paper.md
Larger rewrite
