### 2024-06-12 34a9110
Update paper.md
Final rewrite Tim
