+++ paper/paper.md
@@ -33,7 +33,7 @@ bibliography: paper.bib
 
 Despite robust platforms, specialized technical knowledge is required to setup and run various Machine (ML) and Deep Learning (DL) models, leading many practitioners, scientists, and domain experts to find it difficult to implement them. The **`servir-aces`** Python package is created to fill this gap. **`servir-aces`** significantly lowers the barrier for users to export training data and both train and run DL models using cloud-based technology with their GEE workflows. Several examples are provided via a runnable notebook to make it easier for scientists utilize this emerging field of DL.
 
-With petabytes of data available via GEE, and integration of the TensorFlow (TF) platfrom, models trained in TF can be easily loaded into GEE. This package provides functionalities for 1) data processing; 2) data loading from GEE; 3) feature extraction, 4) model training, and 5) model inference. The combination of TF and GEE has enabled several large scale ML and DL Remote Sensing applications. Some of them include Wetland Area Mapping [@Bakkestuen2023],, Crop Type Mapping [@Bakkestuen2023],, Surface Water Mapping [@Mayer2021], and Urban Mapping [@Parekh2021]. However, these applications tend to be developed ad-hoc without using a common library and require a very specialized domain as well as technical knowledge. In addition, several unified libraries like torchgeo [@Stewart_TorchGeo_Deep_Learning_2022] and [rastervision](https://github.com/azavea/raster-vision) exists but those are mostly targeted for pytorch user community. Some efforts for GEE & TensorFlow users stemmed out of geemap [@Wu2020] but that is mostly for classical ML approaches such as tree based approaches like Random Forest, while [geospatial-ml](https://github.com/opengeos/geospatial-ml) has not seen much development since its inception. Thus, there is a need for a unified libraries to train DL models within the GEE & TensorFlow user community. **`servir-aces`** reduces the barriers for users by 1) offering a streamlined application of commonly employed architectures (CNN, DNN, and U-NET); 2) allowing for end-users to rapidly adjust a wide range of model parameters for these common architectures including (activation functions, optimizers, loss functions, early stopping, dropout rate, batch size etc.); 3) More efficiently and effectively connect across the Google Cloud ecosystem linking between the Google Cloud, improved methods of parallelization via Apache beam,  Vertex AI, TensorFlow, and Google Earth Engine; 4) enabling broader development and incorporation of cutting edge methods through the packages utility function; 5) finally the package offers a collated set of evaluation metrics for easier model performance comparisons, essential for the scientific community. Although **`servir-aces`** was originally developed for agricultural related applications, the library has been further developed to work for any kind of DL image segmentation tasks.
[…]
