+++ paper/paper.bib
@@ -39,6 +39,16 @@
   note={under review}
 }
 
+@misc{bhandari2024comparing,
+  title={Comparing Deep Learning Models for Rice Mapping in Bhutan Using High Resolution Satellite Imagery},
+  author={Biplov Bhandari and Timothy Mayer},
+  year={2024},
+  eprint={2406.07482},
+  archivePrefix={arXiv},
+  primaryClass={cs.CV},
+  url={https://arxiv.org/abs/2406.07482},
+}
+
 @article{Bakkestuen2023,
   title={Delineation of Wetland Areas in South Norway from Sentinel-2 Imagery and LiDAR Using TensorFlow, U-Net, and Google Earth Engine},
   author={Bakkestuen, Vegar and Venter, Zander and Ganer{\o}d, Alexandra Jarna and Framstad, Erik},
