### 2024-06-20 SERVIR/servir-aces#37 @jtagusari
@biplovbhandari
Thank you for revising the manuscript.
On the sample training program, I encountered an error and have not yet finished the training.

The python code is as follow (not changed):
```{python}
from aces.config import Config
from aces.model_trainer import ModelTrainer

config_file = "config.env"
config = Config(config_file)
trainer = ModelTrainer(config)
trainer.train_model()
```

And the output is as follows:
```{python}
{
	"name": "InvalidArgumentError",
	"message": "Expected 'tf.Tensor(False, shape=(), dtype=bool)' to be true. Summarized data: b'No files matched pattern: aces\\\\data\\\\training/*'",
	"stack": "---------------------------------------------------------------------------
InvalidArgumentError                      Traceback (most recent call last)
File c:\\Users\\jtagu\\Desktop\\review-aces\\test2.qmd:7
      5 config = Config(config_file)
      6 trainer = ModelTrainer(config)
----> 7 trainer.train_model()

File c:\\Users\\jtagu\\Desktop\\review-aces\\.venv\\lib\\site-packages\\aces\\model_trainer.py:102, in ModelTrainer.train_model(self)
    100 print(\"****************************************************************************\")
    101 print(\"****************************** creating datasets... ************************\")
--> 102 self.create_datasets(print_info=self.config.PRINT_INFO)
    104 if self.config.USE_AI_PLATFORM:
    105     print(\"****************************************************************************\")

File c:\\Users\\jtagu\\Desktop\\review-aces\\.venv\\lib\\site-packages\\aces\\model_trainer.py:180, in ModelTrainer.create_datasets(self, print_info)
    171 def create_datasets(self, print_info: bool = False) -> None:
    172     \"\"\"
    173     Create TensorFlow datasets for training, testing, and validation.
    174
   (...)
    178     Prints information about the created datasets if print_info is set to True.
    179     \"\"\"
--> 180     self.TRAINING_DATASET = DataProcessor.get_dataset(
    181         f\"{str(self.config.TRAINING_DIR)}/*\",
    182         self.config.FEATURES,
    183         self.config.LABELS,
    184         self.config.PATCH_SHAPE[0],
    185         self.config.BATCH_SIZE,
    186         self.config.OUT_CLASS_NUM,
    187         **{**self.config.__dict__, \"training\": True},
    188     ).repeat()
    190     self.VALIDATION_DATASET = DataProcessor.get_dataset(
    191         f\"{str(self.config.VALIDATION_DIR)}/*\",
    192         self.config.FEATURES,
   (...)
    197         **self.config.__dict__,
    198     ).repeat()
    200     self.TESTING_DATASET = DataProcessor.get_dataset(
    201         f\"{str(self.config.TESTING_DIR)}/*\",
    202         self.config.FEATURES,
   (...)
    207         **self.config.__dict__,
    208     )

File c:\\Users\\jtagu\\Desktop\\review-aces\\.venv\\lib\\site-packages\\aces\\data_processor.py:517, in DataProcessor.get_dataset(pattern, features, labels, patch_size, batch_size, n_classes, **kwargs)
    488 \"\"\"
    489 Get a TFRecord dataset.
    490
   (...)
    513 * tf.data.Dataset: The TFRecord dataset.
    514 \"\"\"
    515 print(f\"Loading dataset from {pattern}\")
--> 517 dataset = tf.data.Dataset.list_files(pattern).interleave(DataProcessor.create_tfrecord_from_file)
    519 if kwargs.get(\"IS_DNN\", False):
    520     if kwargs.get(\"USE_AI_PLATFORM\", False):

File c:\\Users\\jtagu\\Desktop\\review-aces\\.venv\\lib\\site-packages\\tensorflow\\python\\data\\ops\\dataset_ops.py:1320, in DatasetV2.list_files(file_pattern, shuffle, seed, name)
   1313 condition = math_ops.greater(array_ops.shape(matching_files)[0], 0,
   1314                              name=\"match_not_empty\")
   1316 message = math_ops.add(
   1317     \"No files matched pattern: \",
   1318     string_ops.reduce_join(file_pattern, separator=\", \"), name=\"message\")
-> 1320 assert_not_empty = control_flow_assert.Assert(
   1321     condition, [message], summarize=1, name=\"assert_not_empty\")
   1322 with ops.control_dependencies([assert_not_empty]):
   1323   matching_files = array_ops.identity(matching_files)

File c:\\Users\\jtagu\\Desktop\\review-aces\\.venv\\lib\\site-packages\\tensorflow\\python\\util\\traceback_utils.py:153, in filter_traceback.<locals>.error_handler(*args, **kwargs)
    151 except Exception as e:
    152   filtered_tb = _process_traceback_frames(e.__traceback__)
--> 153   raise e.with_traceback(filtered_tb) from None
    154 finally:
    155   del filtered_tb

File c:\\Users\\jtagu\\Desktop\\review-aces\\.venv\\lib\\site-packages\\tensorflow\\python\\ops\\control_flow_assert.py:102, in Assert(condition, data, summarize, name)
    100     xs = ops.convert_n_to_tensor(data)
    101     data_str = [_summarize_eager(x, summarize) for x in xs]
--> 102     raise errors.InvalidArgumentError(
    103         node_def=None,
    104         op=None,
    105         message=\"Expected '%s' to be true. Summarized data: %s\" %
    106         (condition, \"\
\".join(data_str)))
    107   return
    109 with ops.name_scope(name, \"Assert\", [condition, data]) as name:

InvalidArgumentError: Expected 'tf.Tensor(False, shape=(), dtype=bool)' to be true. Summarized data: b'No files matched pattern: aces\\\\data\\\\training/*'"
}
```

I may have the wrong parameters, but I am of the opinion that you should have a look at your documentation. I am posting other issues regarding this.

Thanks,
