### 2024-06-20 SERVIR/servir-aces#37 @jtagusari
@biplovbhandari I tried to follow README instruction on my PC but it fails.

1. sample data was retrieved (dnn data, using gsutil)
2. I set `DATADIR = "DATADIR"` and `MODEL_TYPE = "dnn"`
3. I tried to run the code above, but it fails
