### 2024-06-20 SERVIR/servir-aces#37 @biplovbhandari
@jtagusari The error looks like the files were not found. Can you double check that you're setting the right path to DATADIR? If that doesn't work, I will try to make a video explaining the processes.
