### 2024-06-20 SERVIR/servir-aces#37 @jtagusari
@biplovbhandari
I found that the program was started after I also set `BASEDIR = ""` (please see the [config.zip](https://github.com/user-attachments/files/15912705/config.zip)).

However, another error was thrown (please note that I changed `ipython` to `python` environment to run the program):
```
************************ preparing output directory... *********************
> Saving models and results at output\unet_v1...
Traceback (most recent call last):
  File "c:\Users\jtagu\Desktop\review-aces\test.py", line 7, in <module>
    trainer.train_model()
  File "C:\Users\jtagu\Desktop\review-aces\.venv\lib\site-packages\aces\model_trainer.py", line 115, in train_model
    self.prepare_output_dir()
  File "C:\Users\jtagu\Desktop\review-aces\.venv\lib\site-packages\aces\model_trainer.py", line 154, in prepare_output_dir
    os.mkdir(self.config.MODEL_SAVE_DIR)
FileNotFoundError: [WinError 3] 指定されたパスが見つかりません。: 'output\\unet_v1'
```

Although this error should be fixed, it was understandable, so I created `output/unet_v1` folder and tried to run the program.

However, it throws another error:
```
****************************** training model... ***************************
Traceback (most recent call last):
  File "c:\Users\jtagu\Desktop\review-aces\test.py", line 7, in <module>
    trainer.train_model()
  File "C:\Users\jtagu\Desktop\review-aces\.venv\lib\site-packages\aces\model_trainer.py", line 118, in train_model
    self.start_training()
  File "C:\Users\jtagu\Desktop\review-aces\.venv\lib\site-packages\aces\model_trainer.py", line 252, in start_training
    model_checkpoint = callbacks.ModelCheckpoint(
  File "C:\Users\jtagu\Desktop\review-aces\.venv\lib\site-packages\keras\src\callbacks\model_checkpoint.py", line 191, in __init__
    raise ValueError(
ValueError: The filepath provided must end in `.keras` (Keras model format). Received: filepath=output\unet_v1/modelCheckpoint
```

Please let me know a solution if I misunderstood the instructions. If it was due to the bug of the program, please fix it. And I hope you could prepare a fixed sample program and data that can be executed without changing parameters (As I also posted as another issue).

Thanks.
