### 2024-06-23 f353a03
address issue #46 and #47 improve README bugfixes associated with it
