### 2024-06-23 0931bc3
address issue #46 and #47 improve README bugfixes associated with it
