+++ aces/config.py
@@ -10,6 +10,8 @@ import os
 import ast
 from dotenv import load_dotenv
 
+from aces.utils import Utils
+
 
 class Config:
     """
@@ -30,8 +32,6 @@ class Config:
         FEATURES (str): The list of features used in the model.
         USE_ELEVATION (bool): Flag to use elevation data.
         USE_S1 (bool): Flag to use Sentinel-1 data.
-        DERIVE_FEATURES (bool): Flag to derive features.
-        ADDED_FEATURES (list): A list of additional features used in the model (used when DERIVE_FEATURES is true).
         LABELS (list): A list of labels used in the model.
         SCALE (int): The scale of the data.
         # Note: The seeding componenet needs fixing. It is not working as expected.
@@ -117,19 +117,28 @@ class Config:
             self.DATADIR = self.BASEDIR / os.getenv("DATADIR")
             self.TRAINING_DIR = self.DATADIR / "training"
             self.TESTING_DIR = self.DATADIR / "testing"
-            self.VALIDATION_DIR= self.DATADIR / "validation"
+            self.VALIDATION_DIR = self.DATADIR / "validation"
 
         print(f"BASEDIR: {self.BASEDIR}")
-        print(f"DATADIR: {self.DATADIR}")
+        print(f"checking if the DATADIR exists at: {self.DATADIR}")
+        if not Path(f"{self.DATADIR}").is_dir():
+            raise FileNotFoundError(f"The directory '{self.DATADIR}' does not exist.")
 
         self.OUTPUT_DIR = self.BASEDIR / os.getenv("OUTPUT_DIR")
 
         self.MODEL_NAME = os.getenv("MODEL_NAME")
+        if self.MODEL_NAME is None:
+            self.MODEL_NAME = "aces_model"
+
         self.MODEL_CHECKPOINT_NAME = os.getenv("MODEL_CHECKPOINT_NAME")
 
         self.MODEL_DIR_NAME = os.getenv("MODEL_DIR_NAME")
 
         self.MODEL_DIR = self.OUTPUT_DIR / self.MODEL_DIR_NAME
+        # Ensure MODEL_DIR directory exists, create if it doesn't
+        # this takes care of creating OUTPUT_DIR as well
+        print(f"creating if MODEL_DIR does not exist at: {self.MODEL_DIR}")
+        self.MODEL_DIR.mkdir(parents=True, exist_ok=True)
 
         self.AUTO_MODEL_DIR_NAME = os.getenv("AUTO_MODEL_DIR_NAME") == "True"
 
@@ -137,11 +146,14 @@ class Config:
 
         self.SCALE = int(os.getenv("SCALE"))
 
-        self.FEATURES = os.getenv("FEATURES").split("\n")
-        self.ADDED_FEATURES = os.getenv("ADDED_FEATURES").split("\n")
+        # self.FEATURES = os.getenv("FEATURES").split("\n")
+        self.FEATURES = Utils.parse_params("FEATURES")
+
         self.USE_ELEVATION = os.getenv("USE_ELEVATION") == "True"
         self.USE_S1 = os.getenv("USE_S1") == "True"
-        self.LABELS = os.getenv("LABELS").split("\n")
+
+        # self.LABELS = os.getenv("LABELS").split("\n")
+        self.LABELS = Utils.parse_params("LABELS")
 
         if self.USE_ELEVATION:
                 self.FEATURES.extend(["elevation", "slope"])
@@ -154,15 +166,20 @@ class Config:
[…]
