### 2024-06-23 SERVIR/servir-aces#37 @biplovbhandari
@jtagusari I have fixed this issue. Please try with the newer pip version. I have also updated the [`README.md`](https://github.com/SERVIR/servir-aces/blob/main/README.md) with [Installation](https://github.com/SERVIR/servir-aces?tab=readme-ov-file#installation), [Usage](https://github.com/SERVIR/servir-aces?tab=readme-ov-file#usage), [Tutorial](https://github.com/SERVIR/servir-aces?tab=readme-ov-file#tutorial), etc. This should also address the two other issues that you have opened. Let me know how it goes. Thanks!
