### 2024-06-28 SERVIR/servir-aces#37 @jtagusari
@biplovbhandari Thank you for revising the document. The sections you have write are easy to understand. The meaning of the terms and instructions are clear. In my environment, the tutorial codes look to be working successfully.

I found a few things that may need to be corrected:
1. `mkdir DATADIR` -> `mkdir path/to/your/project/data`
2. `IMGDIR` -> `IMAGEDIR`

And I would like to ask you how to visualize and verify the results. The Tutorial instructs to employ Google, but it requires a billing account. It also introduces a `note` where the same code is provided and executed, but there are no visualized output of the results and cannot tell whether the output is correct. I hope you would let know me the way to do this. Thanks.
