### 2024-06-30 e0e784f
improve README to include the visualization portion
