### 2024-06-30 SERVIR/servir-aces#37 @biplovbhandari
@jtagusari Thank you for your comment. I have made those two changes that you pointed out. In addition, I have also updated the relevant section of the [`README.md`](https://github.com/SERVIR/servir-aces?tab=readme-ov-file#uploading-predictions-to-gcp-and-gee) to reflect changes on how you can visualize the result. This should resolve the visualization feature. Let me know how it goes. Thanks!
