+++ index.md
@@ -1,0 +1,449 @@
+# ACES (Agricultural Classification and Estimation Service)
+
+[![image](https://img.shields.io/pypi/v/servir-aces.svg)](https://pypi.python.org/pypi/servir-aces)
+[![Conda Recipe](https://img.shields.io/badge/recipe-servir--aces-green.svg)](https://github.com/conda-forge/servir-aces-feedstock)
+[![image](https://img.shields.io/conda/vn/conda-forge/servir-aces.svg)](https://anaconda.org/conda-forge/servir-aces)
+[![Conda Downloads](https://img.shields.io/conda/dn/conda-forge/servir-aces.svg)](https://anaconda.org/conda-forge/servir-aces)
+
+ACES (Agricultural Classification and Estimation Service) is a Python module for generating training data and training machine learning models for remote sensing applications. It provides functionalities for data processing, data loading from Earth Engine, feature extraction, and model training.
+
+## Installation
+
+```sh
+pip install -U servir-aces
+```
+
+For saving model plot in TensorFlow/Keras, you need `pydot` and Graphviz. To install, `pypot, run:
+
+```sh
+pip install pydot
+```
+
+The installation process for Graphviz varies depending on your operating system.
+
+- **For macOS**:\
+You can install Graphviz using Homebrew. If you don't have Homebrew installed, you can get it from https://brew.sh Once Homebrew is installed, run the following command:
+
+```sh
+brew install graphviz
+```
+
+- **For Windows**:\
+Download the installer from the Graphviz website ([Graphviz Download Page](https://graphviz.gitlab.io/download/)), and follow the installation instructions. Make sure to add the path to the Graphviz bin folder to your system’s PATH environment variable.
+
+- **For Linux (Ubuntu/Debian)**:\
+You can install Graphviz using apt:
+
+```sh
+sudo apt-get install graphviz
+```
+
+## Features
+
+- Data loading and processing from Earth Engine.
+- Generation of training data for machine learning models.
+- Training and evaluation of machine learning models (DNN, CNN, UNET).
+- Inferences of the trained ML/DL models.
+- Support for remote sensing feature extraction.
+- Integration with Apache Beam for data processing.
+
+
+## Usage
+
+### Setup
+ACES relies on environment variables defined in a `.env` file to configure various aspects of the training process. You can find an example `.env` file named [`.env.example`](https://github.com/SERVIR/servir-aces/blob/main/.env.example). Copy this file and rename it to `config.env` in your project directory. Edit the variables within `config.env` to suit your specific needs. You will need to change several environment settings before running. Below are explained some of the environment configuration.
+
+### Environment Configuration
+- **BASEDIR (str)**: The base directory for the project. This is the root directory where all project-related files and subdirectories reside. It serves as the root directory for other folders containing data, outputs, etc.
+
+- **DATADIR (str)**: The directory for your data, and will be used to retrieve data necessary for training experiments. For this you can either specify the Google Cloud Storage (GCS) path starting with `gs://` or the path which can ben either an full absolute path or relative path to `BASEDIR`. This means, for example, the path to your data directory can be either `gs://aces-project/data` or the full absolute path as `/home/aces-project/data` or can be relative path to `BASEDIR`; so if `BASEDIR` is `/home/aces-project` and `DATADIR` is `data`, then the `DATADIR` is constructed as `/home/aces-project/data`. \
[…]
