+++ metrics/metrics.md
@@ -1,0 +1,4 @@
+::: aces.metrics
+    rendering:
+      show_root_heading: true
+      show_source: true
