### 2024-07-04 #6729 @nagellette
@kanishkan91 how strict should we be about automated tests? I don't see any official test, and related test folder is empty though I think that most of the functionality is provided within example notebooks where one can run and test if those actually work.

Apart from this, I think the improvements in text are enough and latest improvements as per @jtagusari 's feedback considerably improved the code and documentation. Especially the suggestion about the smaller test data is really helpful.
