### 2024-07-05 #6729 @kanishkan91
@nagellette Thanks! I mean soimetinmes even if there are no automated tests, there are some tests embodied in the code itself. Idea would be to ensure that there are tests so that when a new developer adds features, the code is not broken.

@biplovbhandari - Could you comment on this? There should be some tests included in the software package I think to ensure stable continued development in the future. Ideally this should be connected to codecov, so that users can check coverage. Also, given that the tests folder is present but empty, I would assume some tests were to be added?
