### 2024-07-05 #6729 @biplovbhandari
@kanishkan91 I agree with @nagellette that the README and the example notebooks are the primary ways to ensure that the code is working correctly and not broken. These notebooks provide practical examples and allow users to run and test the code with small data samples. Thanks to @jtagusari.
To avoid confusion, I will remove the empty test folder that was automatically generated from the templates. This way, it is clear that the example notebooks are intended to serve as the main verification method for the code's functionality.
Thank you for your understanding.
