+++ tests/__init__.py
@@ -1,1 +1,0 @@
-"""Unit test package for aces."""
