+++ .github/workflows/ubuntu.yml
@@ -40,7 +40,4 @@ jobs:
               run: |
                   pip install --user -r requirements.txt
                   pip install .
-            - name: PKG-TEST
-              run: |
-                  python -m unittest discover tests/
 
