### 2024-07-05 #6729 @kanishkan91
@biplovbhandari Thanks. I just took a look at the sample code. Any chance you could just convert these to tests and put them in the test folder? We want to ensure there are automated tests as per JOSS guidelines.

Converting the code to tests seems strightforward. I would recommend you do that. Let me know what you think.

(PS- I agree that the current code is sufficient so I'm not suggesting adding extra tests, just converting the current ones!)
