+++ tests/__init__.py
@@ -1,0 +1,1 @@
+"""Unit test package for aces."""
