+++ .github/workflows/docs.yml
@@ -18,9 +18,12 @@ jobs:
                   python -m pip install --upgrade pip
                   pip install --user --no-cache-dir Cython
                   pip install --user -r requirements.txt
-            # - name: PKG-TEST
-            #   run: |
-            #       python -m unittest discover tests/
+            - name: Install pytest
+              run: |
+                  pip install pytest
+            - name: PKG-TEST
+              run: |
+                  python -v -s tests/
             - run: python -m pip install --upgrade pip
             - run: pip install mkdocs-material mkdocstrings mkdocstrings-python-legacy mkdocs-git-revision-date-plugin mkdocs-jupyter mkdocs-bibtex mkdocs-git-revision-date-localized-plugin
             - run: mkdocs gh-deploy --force
