+++ .github/workflows/docs.yml
@@ -18,6 +18,12 @@ jobs:
                   python -m pip install --upgrade pip
                   pip install --user --no-cache-dir Cython
                   pip install --user -r requirements.txt
+            - name: Install pydot
+              run: |
+                  pip install pydot
+            - name: Install graphviz
+              run: |
+                  sudo apt-get install graphviz
             - name: Install pytest
               run: |
                   pip install pytest
