+++ .github/workflows/docs.yml
@@ -29,7 +29,7 @@ jobs:
                   pip install pytest
             - name: PKG-TEST
               run: |
-                  python -v -s tests/
+                  pytest -v -s tests/
             - run: python -m pip install --upgrade pip
             - run: pip install mkdocs-material mkdocstrings mkdocstrings-python-legacy mkdocs-git-revision-date-plugin mkdocs-jupyter mkdocs-bibtex mkdocs-git-revision-date-localized-plugin
             - run: mkdocs gh-deploy --force
