### 2024-07-07 #6729 @biplovbhandari
@kanishkan91 I have added the tests to make sure the functionalities are working as intended. The GitHub action also makes sure that tests are running as expected. To run the test, you can do `pytest -v -s tests/` I hope the package is ready to be published now!
