### 2024-07-08 SERVIR/servir-aces#37 @jtagusari
@biplovbhandari I appreciate your continuous contribution to the project and improvement of the program and documentation. I have confirmed that the tutorial program works successfully, while an additional command was required in my environment (see below). Please confirm whether this command is necessary (or it may be due to my own problem) and close this issue. Thank you very much!

In my environment, the following command was necessary before `earthengine upload image`
```
earthengine set_project {my-project-name}
```
