### 2024-07-08 #6729 @kanishkan91
@biplovbhandari - We are getting closer! Below are some minor comments on the paper. Kindly correct these.

- [ ] Line 19- "via runnable" should be "via a runnable". Though, it may be worth changing this to "via a user friendly notebook"?
- [ ] Line 29- Add citation for rastervision
- [ ] Line 31- Change " are mostly for" to "are mostly used for"
- [ ] Line 32- "has not seen much development" should rather be reworded to "has not been developed further since its its inception"
- [ ] Line 47- This should be "for development practitioners," basically just add the "s"
- [ ]  Line 51- Would it be possible to add a hyperlink to the notebook?
- [ ] Line 52- Once again make this "crop-mapping related applications"
- [ ] Line 55- This should be " The major high-level functionalities"

After the above are corrected, We want to make sure the archival has the correct metadata that JOSS requires. This includes a title that matches the paper title and a correct author list.

So as next steps for you after the minor corrections:

- [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list matches the people and order in your paper). You may also add the authors' ORCID.
- [ ] Check that software archive exists and title and author list match JOSS paper (or purposefully do not).
- [ ] Please respond with the 1) version and 2) DOI of the archived version here

After this, I can recommend acceptance. Thank you!
