+++ paper/paper.bib
@@ -107,3 +107,10 @@
   title={geemap: A Python package for interactive mapping with Google Earth Engine},
   journal={Journal of Open Source Software}
 }
+
+@misc{rastervision,
+author = {Azavea/Element 84, Robert Cheetham},
+doi = {10.5281/zenodo.8018177},
+title = {Raster Vision: An open source library and framework for deep learning on satellite and aerial imagery (2017-2023).},
+url = {https://github.com/azavea/raster-vision}
+}
