### 2024-07-08 4a20599
Bump version: 0.0.18 → 0.0.19
