+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "servir-aces"
-version = "0.0.18"
+version = "0.0.19"
 dynamic = [
     "dependencies",
 ]
@@ -47,7 +47,7 @@ universal = true
 
 
 [tool.bumpversion]
-current_version = "0.0.18"
+current_version = "0.0.19"
 commit = true
 tag = true
 
