### 2024-07-08 #6729 @biplovbhandari
@kanishkan91 Thank you for the detailed instructions. Please find the needed information:

1) Version v1

2) DOI: [10.5281/zenodo.12687052](https://doi.org/10.5281/zenodo.12687052)

Please let me know if this looks good and if you need any more information. Thanks!
