### 2024-07-08 #6729 @crvernon
@biplovbhandari - Please make the following adjustments to the paper:

- [ ] LINES 11-12:  "...Applications (Mayer et al., 2023), (Bhandari & Mayer, 2024)." should appear as "...Applications (Mayer et al., 2023; Bhandari & Mayer, 2024)."  See other JOSS articles for demonstrations for how to do this in the paper.md file.
- [ ] LINES 22-25:  No need to separate the items with a semi-colon.  Just use a commas in this case like you did after points 3 and 4.
- [ ] LINE 29:  You are missing the data from your Azavea / Element citation.  In fact, the citation itself is incorrect.  The Zenodo record you are pointing to shows it as:

> Lewis Fishgold, Adeel Hassan, Rob Emanuele, James McClain, Simon Kassel, Annie Zhao, jpolchlo, Joe Morrison, Laurens Bakker, James Santucci, Nathan Holeman, rbreslow, Taylor, Bborie Park, Per Liedman, Umang Kalra, Ammar Sani, Andreas Meier, Chris Brown, … Shinya Uryu. (2023). azavea/raster-vision: Raster Vision 0.20.2 (v0.20.2-zenodo). Zenodo. https://doi.org/10.5281/zenodo.8018177

Please adjust this accordingly.  I do know that there is a newer version released, but it does not cite the full author group as this one does.  Change the version number in the above citation if you feel that you need to reference the newer one.

- [ ] LINE 38:  Missing a comma after "...batch size"
- [ ] Throughout this paper:  remote sensing does not need to have the R or S capitalized.  It is not a product in itself.
- [ ] LINES 43-44:  "...essential for scientific community..." is missing a "the" between "for" and "scientific"
- [ ] LINE 54:  "...can be repeated for any kind of the image segmentation task." should be "...can be repeated for any of of image segmentation task."
- [ ] LINES 56-60:  you intended bullet points via the dashes are not working here.
- [ ] Please ensure that capitalization is present where needed in your references.  You can do this by using curly brackets around the content you wish to maintain formatting for in your bib file.  For example, LINE 94:  the "n" in "norway" should be capitalized.  Please check all references.
- [ ] There is a DOI for this record:

> Bakkestuen, V., Venter, Z., Ganerød, A. J., & Framstad, E. (2023). Delineation of wetland areas in south norway from sentinel-2 imagery and LiDAR using TensorFlow, u-net, and google earth engine. Remote Sensing, 15(5), 1203.

Yet there is none listed in your citation.  Please fix this.

- [ ] There are other citations that should have DOI but do not.  Please add these in.
