+++ paper/paper.md
@@ -27,33 +27,34 @@ bibliography: paper.bib
 
 # Summary
 
-**`servir-aces`** Agricultural Classification and Estimation Service (ACES) is a Python package for generating training data using highly parallelized [apache-beam](https://beam.apache.org/) and [Google Earth Engine (GEE)](https://earthengine.google.com/) [@Gorelick2017] workflows as well as for training various Machine Learning (ML) and Deep Learning (DL) models for Remote Sensing Applications [@Mayer2023], [@bhandari2024comparing].
+**`servir-aces`** Agricultural Classification and Estimation Service (ACES) is a Python package for generating training data using highly parallelized [apache-beam](https://beam.apache.org/) and [Google Earth Engine (GEE)](https://earthengine.google.com/) [@Gorelick2017] workflows as well as for training various Machine Learning (ML) and Deep Learning (DL) models for remote sensing Applications [@Mayer2023; @bhandari2024comparing].
 
 # Statement of Need
 
 Despite robust platforms, specialized technical knowledge is required to set up and run various Machine Learning (ML) and Deep Learning (DL) models, making it difficult for many development practitioners, scientists, and domain experts to implement them. The **`servir-aces`** Python package is designed to address this challenge by significantly lowering the barrier for users to export training data and both train and run DL models using cloud-based technology with their GEE workflows. Several examples are provided via a user friendly notebook to make it easier for scientists to utilize this emerging field of DL.
 
-With petabytes of data available via GEE, and integration of the TensorFlow (TF) platfrom, models trained in TF can be easily loaded into GEE. This package provides functionalities for 1) data processing; 2) data loading from GEE; 3) feature extraction, 4) model training, and 5) model inference. The combination of TF and GEE has enabled several large scale ML and DL Remote Sensing applications, including Wetland Area Mapping [@Bakkestuen2023], Crop Type Mapping [@Bakkestuen2023], Surface Water Mapping [@Mayer2021], and Urban Mapping [@Parekh2021]. However, these applications tend to be developed ad-hoc without using a common package.
+With petabytes of data available via GEE, and integration of the TensorFlow (TF) platfrom, models trained in TF can be easily loaded into GEE. This package provides functionalities for 1) data processing, 2) data loading from GEE, 3) feature extraction, 4) model training, and 5) model inference. The combination of TF and GEE has enabled several large scale ML and DL remote sensing applications, including Wetland Area Mapping [@Bakkestuen2023], Crop Type Mapping [@Bakkestuen2023], Surface Water Mapping [@Mayer2021], and Urban Mapping [@Parekh2021]. However, these applications tend to be developed ad-hoc without using a common package.
 
 Several unified libraries like torchgeo [@Stewart_TorchGeo_Deep_Learning_2022] and rastervision [@rastervision] exists, but they are primarily targeted for PyTorch user community. Some efforts for GEE & TensorFlow users, such as geemap [@Wu2020], are mostly used for classical ML approaches like Random Forest, while [geospatial-ml](https://github.com/opengeos/geospatial-ml) has not been developed further since its inception. Thus, there is a need for unified libraries to train DL models integrating the GEE & TensorFlow user community.
 
[…]
