+++ paper/paper.bib
@@ -1,116 +1,120 @@
 @article{Gorelick2017,
-  title={Google Earth Engine: Planetary-scale geospatial analysis for everyone},
-  author={Gorelick, Noel and Hancher, Matt and Dixon, Mike and Ilyushchenko, Simon and Thau, David and Moore, Rebecca},
-  journal={Remote sensing of Environment},
-  volume={202},
-  pages={18--27},
-  year={2017},
-  publisher={Elsevier}
+  title="{Google Earth Engine: Planetary-scale geospatial analysis for everyone}",
+  author="Gorelick, Noel and Hancher, Matt and Dixon, Mike and Ilyushchenko, Simon and Thau, David and Moore, Rebecca",
+  journal="Remote sensing of Environment",
+  volume=202,
+  pages="18--27",
+  year=2017,
+  publisher="Elsevier",
+  doi="10.1016/j.rse.2017.06.031",
 }
 
 @article{Mayer2021,
-  title={Deep learning approach for Sentinel-1 surface water mapping leveraging Google Earth Engine},
-  author={Mayer, Timothy and Poortinga, Ate and Bhandari, Biplov and Nicolau, Andrea P and Markert, Kel and Thwal, Nyein Soe and Markert, Amanda and Haag, Arjen and Kilbride, John and Chishtie, Farrukh and others},
-  journal={ISPRS Open Journal of Photogrammetry and Remote Sensing},
-  volume={2},
-  pages={100005},
-  year={2021},
-  publisher={Elsevier}
+  title="{Deep learning approach for Sentinel-1 surface water mapping leveraging Google Earth Engine}",
+  author="Mayer, Timothy and Poortinga, Ate and Bhandari, Biplov and Nicolau, Andrea P and Markert, Kel and Thwal, Nyein Soe and Markert, Amanda and Haag, Arjen and Kilbride, John and Chishtie, Farrukh and others",
+  journal="ISPRS Open Journal of Photogrammetry and Remote Sensing",
+  volume=2,
+  pages=100005,
+  year=2021,
+  publisher="Elsevier",
+  doi="10.1016/j.ophoto.2021.100005"
 }
 
 @article{Mayer2023,
-  title={Employing the agricultural classification and estimation service (ACES) for mapping smallholder rice farms in Bhutan},
-  author={Mayer, Timothy and Bhandari, Biplov and Mart{\'\i}nez, Filoteo G{\'o}mez and Walker, Kaitlin and Jim{\'e}nez, Stephanie A and Kruskopf, Meryl and Maganini, Micky and Phalke, Aparna and Wangchen, Tshering and Phuntsho, Loday and others},
-  journal={Frontiers in Environmental Science},
-  volume={11},
-  pages={1137835},
-  year={2023},
-  publisher={Frontiers}
+  title="{Employing the agricultural classification and estimation service (ACES) for mapping smallholder rice farms in Bhutan}",
+  author="Mayer, Timothy and Bhandari, Biplov and Mart{\'\i}nez, Filoteo G{\'o}mez and Walker, Kaitlin and Jim{\'e}nez, Stephanie A and Kruskopf, Meryl and Maganini, Micky and Phalke, Aparna and Wangchen, Tshering and Phuntsho, Loday and others",
+  journal="Frontiers in Environmental Science",
+  volume=11,
+  pages=1137835,
+  year=2023,
+  publisher="Frontiers",
+  doi="10.3389/fenvs.2023.1137835"
 }
 
 @article{Bhandari2024,
-  title={Comparing Deep Learning Models for Rice Mapping in Bhutan using High Resolution Satellite Imagery},
-  author={Bhandari, Biplov and Mayer, Timothy},
[…]
