### 2024-07-09 #6729 @crvernon
Thanks @biplovbhandari!

I didn't see the following addressed:

> - [ ]  LINE 29: You are missing the data from your Azavea / Element citation. In fact, the citation itself is incorrect. The Zenodo record you are pointing to shows it as:
>
> > Lewis Fishgold, Adeel Hassan, Rob Emanuele, James McClain, Simon Kassel, Annie Zhao, jpolchlo, Joe Morrison, Laurens Bakker, James Santucci, Nathan Holeman, rbreslow, Taylor, Bborie Park, Per Liedman, Umang Kalra, Ammar Sani, Andreas Meier, Chris Brown, … Shinya Uryu. (2023). azavea/raster-vision: Raster Vision 0.20.2 (v0.20.2-zenodo). Zenodo. https://doi.org/10.5281/zenodo.8018177
>
> Please adjust this accordingly. I do know that there is a newer version released, but it does not cite the full author group as this one does. Change the version number in the above citation if you feel that you need to reference the newer one.

Also, I messed you up on this one.  My apologies.  Please modify the following:

- [ ] LINE 54:  "...be repeated for any of image segmentation task."  should be "...be repeated for any image segmentation task."  Just remove the "of".
