### 2024-07-09 #6729 @biplovbhandari
@crvernon For `raster-vision` I was looking at their [citation file](https://github.com/azavea/raster-vision/blob/master/CITATION.cff) in their [repo](https://github.com/azavea/raster-vision). Maybe I didn't understand fully, should we change to keep all the authors as you pointed above, and keep the [version 0.20.0](https://zenodo.org/records/8018177)?

I will fix the LINE 54 together with this.
