### 2024-07-09 3c447fd
address joss paper comments
