+++ paper/paper.bib
@@ -117,4 +117,5 @@
   url="https://github.com/azavea/raster-vision",
   author="Fishgold, Lewis and Hassan, Adeel and Emanuele, Rob and McClain, James and Kassel, Simon and Zhao, Annie and {jpolchlo} and Morrison, Joe and Bakker, Laurens and Santucci, James and Holeman, Nathan and {rbreslow} and {Taylor} and Park, Bborie and Liedman, Per and Kalra, Umang and Sani, Ammar and Meier, Andreas and Brown, Chris and McCallum, Chuck and Kettler, Cole and Cheipesh, Eugene and {Grigory} and Maleski, Jerome and Bertrand, Matt and Uryu, Shinya",
   doi="10.5281/zenodo.8018177",
+  year=2023,
 }
